moved logout route from app.py to logout.py
