+++ routes/__init__.py
@@ -1,3 +1,4 @@
 from .home import home_bp
 from .register import register_bp
 from .login import login_pb
+from .logout import logout_pb
