+++ templates/profile.html
@@ -268,7 +268,7 @@ fetch("{{ url_for('static', filename='sentiments.json') }}")
 
     <div class="button-container">
         <button class="toggle" onclick="toggleUploadForm()">Click to Upload or Drag & Drop Anywhere</button>
-        <button class="logout-button"><a href="{{ url_for('logout') }}">Logout</a></button>
+        <button class="logout-button"><a href="{{ url_for('logout.logout') }}">Logout</a></button>
     </div>
 
     <div class="toggle-container">
